adds rbac utils to utils folder and updates typedocs
