+++ x-pack/plugins/rule_registry/server/rule_data_client/types.ts
@@ -11,7 +11,7 @@ import { ElasticsearchClient } from 'kibana/server';
 import { FieldDescriptor } from 'src/plugins/data/server';
 import { ESSearchRequest, ESSearchResponse } from 'src/core/types/elasticsearch';
 import { TechnicalRuleDataFieldName } from '../../common/technical_rule_data_field_names';
-import { ValidFeatureId } from '../alert_data_client/alerts_client';
+import { ValidFeatureId } from '../utils/rbac';
 
 export interface RuleDataReader {
   search<TSearchRequest extends ESSearchRequest>(
